Refacto: split utils.
